Add test filtering with -f/--filter option. Verify that special chars can be thrown through stdout/err.

+++ paratec.c
@@ -6,18 +6,23 @@
  * http://opensource.org/licenses/MIT
  */
 
+#define _GNU_SOURCE
 #include <errno.h>
+#include <fcntl.h>
+#include <getopt.h>
+#include <signal.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <sys/mman.h>
 #include <sys/time.h>
 #include <sys/wait.h>
-#include <sys/mman.h>
-#include <signal.h>
 #include <time.h>
 #include <unistd.h>
 #include "paratec.h"
 
+#define INDENT "    "
+#define STDPREFIX INDENT INDENT " | "
 #define LINE_SIZE 2048
 
 #define MIN(a, b) (a < b ? a : b)
@@ -25,22 +30,37 @@
 #define N_ELEMENTS(arr) (sizeof(arr) / sizeof((arr)[0]))
 
 struct job {
-	uint32_t i; // Index of the test in tests.all
 	pid_t pid;
+	int stdout;
+	int stderr;
+
+	uint32_t i; // Index of the test in tests.all
 	int64_t end_at;
 	int timed_out;
 	char last_line[LINE_SIZE];
 };
 
+struct buff {
+	char *str;
+	size_t len;
+	size_t alloc;
+};
+
 struct test {
 	char name[LINE_SIZE];
-	int passed;
-	int timed_out;
 	int exit_status;
 	int signal_num;
-	int64_t i;
+	int64_t i; // For ranged tests, the i in the range
+	struct buff stdout;
+	struct buff stderr;
 	char last_line[LINE_SIZE];
 	struct paratec *p;
+	struct {
+		int run:1;
+		int filtered_run:1;
+		int passed:1;
+		int timed_out:1;
+	} flags;
 };
 
 struct tests {
@@ -50,12 +70,14 @@ struct tests {
 
 	uint32_t c;
 	uint32_t allocated;
+	uint32_t enabled;
+	uint32_t next_test;
 	struct test *all;
 };
 
 static uint32_t _max_jobs;
 static uint32_t _timeout;
-static uint32_t _verbose;
+static int _verbose;
 
 // Only for use in testing processes
 static struct job *_tjob;
@@ -93,7 +115,7 @@ static void _signal_wait()
 	siginfo_t info;
 	struct timespec timeout = {
 		.tv_sec = 0,
-		.tv_nsec = 100 * 1000,
+		.tv_nsec = 10 * 1000,
 	};
 
 	sigemptyset(&mask);
@@ -121,16 +143,74 @@ static void _setup_signals()
 	}
 }
 
-static void _set_opts(int argc, char **argv)
+
+static void _filter_tests(struct tests *ts, const char *filter)
 {
+	uint32_t i;
+
+	for (i = 0; i < ts->c; i++) {
+		struct test *t = ts->all + i;
+		if (strstr(t->name, filter) == t->name) {
+			t->flags.filtered_run = 1;
+		} else {
+			t->flags.run = 0;
+		}
+	}
+}
+
+static void _print_opt(char *s, char *arg, char *desc)
+{
+	printf(INDENT "-%s, --%s\n", s, arg);
+	printf(INDENT INDENT "%s\n", desc);
+}
+
+static void _print_usage(char **argv)
+{
+	printf("Usage: %s [OPTION]...\n", argv[0]);
+	printf("\n");
+	_print_opt("f FILTER", "filter=FILTER", "only run tests prefixed with FILTER");
+	_print_opt("h", "help", "print this messave");
+	_print_opt("v", "verbose", "print information about succeeding tests");
+
+	exit(2);
+}
+
+static void _set_opts(struct tests *ts, int argc, char **argv)
+{
+	struct option lopts[] = {
+		{ "filter", required_argument, NULL, 'f' },
+		{ "help", no_argument, NULL, 'h' },
+		{ "verbose", no_argument, &_verbose, 'v' },
+		{ NULL, 0, NULL, 0 },
+	};
+
 	_max_jobs = _get_cpu_count() + 1;
 	_timeout = 5;
 
-	if (argc == 1) {
-		return;
+	while (1) {
+		char c = getopt_long(argc, argv, "f:hv", lopts, NULL);
+		if (c == -1) {
+			break;
+		}
+
+		switch (c) {
+			case 0:
+				break;
+
+			case 'f':
+				_filter_tests(ts, optarg);
+				break;
+
+			case 'v':
+				_verbose = 1;
+				break;
+
+			case 'h':
+			default:
+				_print_usage(argv);
+		}
 	}
 
-	// @todo test filtering
 	// @todo single-process testing
 	// @todo recording last-executed line of test
 	// @todo global setup/teardown in parent process
@@ -146,8 +226,11 @@ static void _add_test(struct tests *ts, struct paratec *p)
 	int has_range = 0;
 
 	struct test t = {
-		.passed = 0,
 		.p = p,
+		.flags = {
+			.run = 1,
+			.passed = 0,
+		},
 	};
 
 	if (p->range_low == 0 && p->range_high == 0) {
@@ -160,6 +243,10 @@ static void _add_test(struct tests *ts, struct paratec *p)
 		if (ts->c == ts->allocated) {
 			ts->allocated = MAX(ts->allocated, 1) * 2;
 			ts->all = realloc(ts->all, sizeof(*ts->all) * ts->allocated);
+			if (ts->all == NULL) {
+				fprintf(stderr, "failed to allocate array for tests\n");
+				exit(1);
+			}
 		}
 
 		if (has_range) {
@@ -169,9 +256,54 @@ static void _add_test(struct tests *ts, struct paratec *p)
 		}
 
 		t.i = i;
-		snprintf(t.name, sizeof(t.name), "%s:%s%s", p->file, p->name, index);
+		snprintf(t.name, sizeof(t.name), "%s%s", p->name, index);
 
 		ts->all[ts->c++] = t;
+	}
+}
+
+static void _flush_pipe(int fd, struct buff *b)
+{
+	ssize_t err;
+	ssize_t cap;
+
+	do {
+		if (b->alloc == b->len) {
+			b->alloc = sizeof(*b->str) * (MAX(b->alloc, 16) * 2);
+			b->str = realloc(b->str, b->alloc);
+			if (b->str == NULL) {
+				fprintf(stderr, "failed to allocate buffer for test output\n");
+				exit(1);
+			}
+		}
+
+		cap = (b->alloc - b->len) - 1;
+		err = read(fd, b->str + b->len, cap);
+		if (err < 0) {
+			if (errno != EAGAIN && errno != EWOULDBLOCK) {
+				perror("failed to read test output");
+				exit(1);
+			}
+		} else {
+			b->len += err;
+			b->str[b->len] = '\0';
+		}
+	} while (err == cap);
+}
+
+static void _flush_pipes(struct tests *ts, struct job *jobs, uint32_t jobsc)
+{
+	uint32_t i;
+
+	for (i = 0; i < jobsc; i++) {
+		struct job *j = jobs + i;
+
+		if (j->pid == -1) {
+			continue;
+		}
+
+		_flush_pipe(j->stdout, &ts->all[j->i].stdout);
+		_flush_pipe(j->stderr, &ts->all[j->i].stdout);
 	}
 }
 
@@ -186,7 +318,7 @@ static void _check_timeouts(struct job *jobs, uint32_t jobsc)
 		if (j->pid != -1 && j->end_at < now) {
 			j->timed_out = 1;
 			err = killpg(j->pid, SIGKILL);
-			if (err < 0) {
+			if (err < 0 && errno != ESRCH) {
 				perror("failed to kill child test");
 				exit(1);
 			}
@@ -194,12 +326,43 @@ static void _check_timeouts(struct job *jobs, uint32_t jobsc)
 	}
 }
 
+static void _pipe(int p[2])
+{
+	int err = pipe(p);
+	if (err < 0) {
+		perror("failed to create pipes");
+		exit(1);
+	}
+
+	err = fcntl(p[0], F_SETFL, O_NONBLOCK);
+	if (err < 0) {
+		perror("failed to make test pipe nonblocking");
+		exit(1);
+	}
+}
+
+static void _dup2(int std, int fd)
+{
+	int err = dup2(fd, std);
+	if (err < 0) {
+		perror("failed to dup2");
+		exit(1);
+	}
+}
+
 static void _run_test(struct test *t, struct job *j)
 {
 	int err;
 	pid_t pid;
+	int pstdin[2];
+	int pstdout[2];
+	int pstderr[2];
 
 	memset(j, 0, sizeof(*j));
+
+	_pipe(pstdin);
+	_pipe(pstdout);
+	_pipe(pstderr);
 
 	pid = fork();
 	if (pid == -1) {
@@ -210,6 +373,10 @@ static void _run_test(struct test *t, struct job *j)
 	}
 
 	if (pid == 0) {
+		_dup2(STDIN_FILENO, pstdin[1]);
+		_dup2(STDOUT_FILENO, pstdout[1]);
+		_dup2(STDERR_FILENO, pstderr[1]);
+
 		err = setpgid(0, 0);
 		if (err < 0) {
 			perror("could not setpgid");
@@ -231,8 +398,26 @@ static void _run_test(struct test *t, struct job *j)
 
 		exit(0);
 	} else {
-		j->end_at = _now() + ((t->p->timeout ?: _timeout) * 1000 * 1000);
+		close(pstdin[0]);
+
+		j->stdout = pstdout[0];
+		j->stderr = pstderr[0];
 		j->pid = pid;
+		j->end_at = _now() + ((t->p->timeout ?: _timeout) * 1000 * 1000);
+	}
+}
+
+static void _run_next_test(struct tests *ts, struct job *j)
+{
+	while (ts->next_test < ts->c) {
+		uint32_t i = ts->next_test++;
+		struct test *t = ts->all + i;
+
+		if (t->flags.run) {
+			_run_test(t, j);
+			j->i = i;
+			break;
+		}
 	}
 }
 
@@ -246,14 +431,13 @@ static void _run_tests(struct tests *ts)
 	struct job *jobsmm;
 	struct job jobs[_max_jobs];
 	uint32_t finished = 0;
-	uint32_t curr_test = 0;
 
 	if (ts->c == 0) {
 		return;
 	}
 
 	jobsmm = mmap(
-		jobs, sizeof(jobs),
+		NULL, sizeof(jobs),
 		PROT_READ | PROT_WRITE,
 		MAP_ANON | MAP_SHARED, -1, 0);
 	if (jobsmm == MAP_FAILED) {
@@ -267,13 +451,10 @@ static void _run_tests(struct tests *ts)
 
 	for (i = 0; i < N_ELEMENTS(jobs) && i < ts->c; i++) {
 		j = jobsmm + i;
-
-		_run_test(ts->all + i, j);
-		j->i = i;
-		curr_test++;
+		_run_next_test(ts, j);
 	}
 
-	while (finished < ts->c) {
+	while (finished < ts->enabled) {
 		struct test *t = NULL;
 
 		_signal_wait();
@@ -288,8 +469,7 @@ static void _run_tests(struct tests *ts)
 			for (i = 0; i < N_ELEMENTS(jobs); i++) {
 				j = jobsmm + i;
 				if (j->pid == pid) {
-					t = ts->all + jobsmm[i].i;
-					j->pid = -1;
+					t = ts->all + j->i;
 					break;
 				}
 			}
@@ -303,22 +483,30 @@ static void _run_tests(struct tests *ts)
 
 			if (WIFEXITED(status)) {
 				t->exit_status = WEXITSTATUS(status);
-				t->passed = t->exit_status == t->p->exit_status;
+				t->flags.passed = t->exit_status == t->p->exit_status;
 			} else if (WIFSIGNALED(status)) {
 				t->signal_num = WTERMSIG(status);
-				t->passed = t->signal_num == t->p->signal_num;
+				t->flags.passed = t->signal_num == t->p->signal_num;
 			} else {
 				continue;
 			}
 
+			_flush_pipes(ts, jobsmm, N_ELEMENTS(jobs));
+
+			close(j->stdout);
+			close(j->stderr);
+			j->pid = -1;
+			j->stdout = -1;
+			j->stderr = -1;
+
 			strncpy(t->last_line, j->last_line, sizeof(t->last_line));
 
-			if (t->passed) {
+			if (t->flags.passed) {
 				ts->passes++;
 				printf(".");
 			} else if (t->signal_num > 0) {
 				ts->errors++;
-				t->timed_out = j->timed_out;
+				t->flags.timed_out = j->timed_out;
 				printf("E");
 			} else {
 				ts->failures++;
@@ -327,19 +515,56 @@ static void _run_tests(struct tests *ts)
 
 			fflush(stdout);
 
-			if (curr_test < ts->c) {
-				_run_test(ts->all + curr_test, j);
-				j->i = curr_test;
-				curr_test++;
-			}
-
+			_run_next_test(ts, j);
 			finished++;
 		}
 
+		_flush_pipes(ts, jobsmm, N_ELEMENTS(jobs));
 		_check_timeouts(jobsmm, N_ELEMENTS(jobs));
 	}
 
 	printf("\n");
+}
+
+static void _dump_line(const char *line, const size_t len)
+{
+	fwrite(STDPREFIX, strlen(STDPREFIX), 1, stdout);
+	fwrite(line, len, 1, stdout);
+	fwrite("\n", 1, 1, stdout);
+}
+
+static void _clear_buff(int dump, const char *which, struct buff *b)
+{
+	if (dump) {
+		printf(INDENT INDENT "%s\n", which);
+
+		if (b->len == 0) {
+			printf(INDENT INDENT " | (nothing)\n");
+		} else {
+			int first = 1;
+			char *start = b->str;
+			size_t len = b->len;
+
+			while (1) {
+				char *nl = memmem(start, len, "\n", 1);
+				if (nl == NULL) {
+					if (first) {
+						_dump_line(start, len);
+					}
+					break;
+				}
+
+				first = 0;
+				*nl = '\0';
+				_dump_line(start, nl - start);
+				len -= nl - start;
+				start = nl + 1;
+			}
+		}
+	}
+
+	free(b->str);
+	b->str = NULL;
 }
 
 int main(int argc, char **argv)
@@ -350,7 +575,6 @@ int main(int argc, char **argv)
 	memset(&ts, 0, sizeof(ts));
 
 	_setup_signals();
-	_set_opts(argc, argv);
 
 	/**
 	 * There's something weird going on with GCC/LD: when not using a pointer,
@@ -367,40 +591,65 @@ int main(int argc, char **argv)
 		sp++;
 	}
 
-	_run_tests(&ts);
-
-	printf("%d%%: %u tests, %u errors, %u failures\n",
-		(int)((((double)ts.passes) / ts.c) * 100),
-		ts.c,
-		ts.errors,
-		ts.failures);
+	_set_opts(&ts, argc, argv);
 
 	for (i = 0; i < ts.c; i++) {
 		struct test *t = ts.all + i;
+		if (t->flags.run || t->flags.filtered_run) {
+			t->flags.run = 1;
+			ts.enabled++;
+		}
+	}
 
-		if (t->passed) {
+	_run_tests(&ts);
+
+	printf("%d%%: %u tests, %u errors, %u failures, %u skipped\n",
+		ts.enabled == 0 ?
+			100 :
+			(int)((((double)ts.passes) / ts.enabled) * 100),
+		ts.c,
+		ts.errors,
+		ts.failures,
+		ts.c - ts.enabled);
+
+	for (i = 0; i < ts.c; i++) {
+		int dump = 1;
+		struct test *t = ts.all + i;
+
+		if (!t->flags.run) {
+			dump = 0;
 			if (_verbose) {
-				printf("\t PASS : %s \n",
+				printf(INDENT " SKIP : %s \n",
+					t->name);
+			}
+		} else if (t->flags.passed) {
+			dump = _verbose;
+			if (dump) {
+				printf(INDENT " PASS : %s \n",
 					t->name);
 			}
 		} else if (t->exit_status != 0) {
-			printf("\t FAIL : %s : exit code=%d\n",
+			printf(INDENT " FAIL : %s : exit code=%d\n",
 				t->name,
 				t->exit_status);
-		} else if (t->timed_out) {
-			printf("\tERROR : %s : timed out after %s\n",
+		} else if (t->flags.timed_out) {
+			printf(INDENT "ERROR : %s : timed out after %s\n",
 				t->name,
 				t->last_line);
 		} else {
-			printf("\tERROR : %s : received signal(%d) `%s` after %s\n",
+			printf(INDENT "ERROR : %s : received signal(%d) `%s` after %s\n",
 				t->name,
 				t->signal_num,
 				strsignal(t->signal_num),
 				t->last_line);
 		}
+
+		_clear_buff(dump, "stdout", &t->stdout);
+		_clear_buff(dump, "stderr", &t->stderr);
 	}
 
 	free(ts.all);
+	ts.all = NULL;
 
-	return ts.passes != ts.c;
+	return ts.passes != ts.enabled;
 }
